Refactor: rename folder, rename method
